% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_button}
\alias{action_button}
\title{Action Button Input}
\usage{
action_button(
  inputId,
  label,
  icon = NULL,
  iconPos = "right",
  disabled = FALSE,
  rounded = FALSE,
  text = FALSE,
  raised = FALSE,
  outlined = FALSE,
  size = "normal"
)
}
\description{
Action Button Input for Shiny from PrimeReact module
}
