% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_flags}
\alias{country_flags}
\title{Available Country Flags}
\format{
A data frame with 22 rows and 3 columns:
\describe{
\item{code}{Country code (lowercase) - use as \code{item} value}
\item{name}{Country name - use as \code{title} value}
\item{code_upper}{Country code (uppercase) - alternative \code{item} value}
}
}
\usage{
country_flags
}
\description{
A data frame containing country codes and names for the embedded country flag icons.
Use these codes with \code{iconClass = "country-flag country-flag-"} in select_input.
}
\examples{
# View available countries
primereact::country_flags

# Use in select_input
\dontrun{
select_input(
  inputId = "country",
  options = lapply(1:nrow(country_flags), function(i) {
    list(title = country_flags$name[i], item = country_flags$code_upper[i])
  }),
  iconClass = "country-flag country-flag-"
)
}

}
\keyword{datasets}
