% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_select.R
\name{extract_items_prime_react_multiselect}
\alias{extract_items_prime_react_multiselect}
\title{Extract Multi-Select Input}
\usage{
extract_items_prime_react_multiselect(named_vec, item = "item")
}
\description{
Extract Multi-Select Input
}
