% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_input}
\alias{select_input}
\title{Select Input}
\usage{
select_input(
  inputId,
  value = NULL,
  options,
  placeholder = "Select value",
  class = NULL,
  width = NULL,
  iconClass = NULL
)
}
\description{
Select Input
}
