% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_button.R
\name{toggle_button}
\alias{toggle_button}
\title{ToggleButton}
\usage{
toggle_button(
  inputId,
  value = FALSE,
  onLabel = "On",
  offLabel = "Off",
  onIcon = "pi pi-check",
  offIcon = "pi pi-times",
  class = NULL,
  width = NULL,
  disabled = FALSE
)
}
\description{
ToggleButton
}
