% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_input.R
\name{update_date_input}
\alias{update_date_input}
\title{Update Date Input}
\usage{
update_date_input(
  session,
  inputId,
  value = NULL,
  placeholder = NULL,
  class = NULL,
  minDate = NULL,
  maxDate = NULL,
  width = NULL,
  readonly = NULL,
  dateFormat = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID}

\item{value}{New date value (optional)}

\item{placeholder}{New placeholder text (optional)}

\item{class}{New CSS class (optional)}

\item{minDate}{New minimum date (optional)}

\item{maxDate}{New maximum date (optional)}

\item{width}{New width (optional)}

\item{readonly}{New readonly setting (optional)}

\item{dateFormat}{New date format (optional)}
}
\description{
Update Date Input. Only pass the parameters you want to update.
}
