% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{use_country_flags}
\alias{use_country_flags}
\title{Country Flags Icons (Auto-loaded)}
\usage{
use_country_flags()
}
\value{
NULL (invisibly). Icons are auto-loaded with the package.
}
\description{
Country flag icons are automatically embedded in the primereact package.
No need to call this function - icons are bundled and load automatically.
}
\details{
Use with select_input or multiple_select_input by setting:
iconClass = "country-flag country-flag-"
}
\examples{
\dontrun{
ui <- fluidPage(
  # No need to call use_country_flags() - auto-loaded!
  select_input(
    inputId = "country",
    value = NULL,
    options = list(
      list(title = "Kenya", item = "KEN"),
      list(title = "Indonesia", item = "IDN")
    ),
    iconClass = "country-flag country-flag-"
  )
)
}

}
