% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_button}
\alias{action_button}
\title{Action Button Input}
\usage{
action_button(
  inputId,
  label,
  translation_list,
  icon = NULL,
  iconPos = "right",
  disabled = FALSE,
  rounded = FALSE,
  text = FALSE,
  raised = FALSE,
  outlined = FALSE,
  size = "normal",
  default_langauge = "en",
  message_handler_id_from_shiny = "language_changed"
)
}
\description{
Action Button Input for Shiny from PrimeReact module
}
