% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{date_range_input}
\alias{date_range_input}
\title{Date Range Input}
\usage{
date_range_input(
  inputId,
  value,
  placeholder = NULL,
  class = NULL,
  minDate = NULL,
  maxDate = NULL,
  width = NULL
)
}
\description{
Date Range Input
}
