% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_select.R
\name{multiple_select_input}
\alias{multiple_select_input}
\title{Multiple Select Input}
\usage{
multiple_select_input(
  inputId,
  value = NULL,
  options,
  placeholder = "Select option(s)",
  class = NULL,
  width = NULL,
  iconClass = NULL
)
}
\description{
Multiple Select Input
}
