% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_select.R
\name{multiple_select_input}
\alias{multiple_select_input}
\title{Multiple Select Input}
\usage{
multiple_select_input(
  inputId,
  value = NULL,
  options,
  placeholder = "Select option(s)",
  class = NULL,
  width = NULL,
  iconClass = NULL,
  filter = FALSE,
  filterInputAutoFocus = FALSE,
  translation_list,
  default_langauge = "en",
  message_handler_id_from_shiny = "language_changed"
)
}
\description{
Multiple Select Input
}
