% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_input}
\alias{select_input}
\title{Select Input}
\usage{
select_input(
  inputId,
  value = NULL,
  options,
  placeholder = "Select value",
  class = NULL,
  width = NULL,
  iconClass = NULL,
  translation_list,
  default_langauge = "en",
  message_handler_id_from_shiny = "language_changed"
)
}
\description{
Select Input
}
