% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_text_button.R
\name{toggle_text_button}
\alias{toggle_text_button}
\title{ToggleTextButton}
\usage{
toggle_text_button(
  inputId,
  value = list(value = NULL, name = NULL),
  options,
  default_langauge = "en",
  translation_list,
  message_handler_id_from_shiny = "language_changed"
)
}
\description{
ToggleTextButton
}
