% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{update_action_button}
\alias{update_action_button}
\title{Update Action Button Input}
\usage{
update_action_button(session, inputId, configuration = NULL)
}
\description{
Update Action Button Input for Shiny from PrimeReact module
}
