% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_select.R
\name{update_multiple_select_input}
\alias{update_multiple_select_input}
\title{Update a multiple_select input}
\usage{
update_multiple_select_input(session, inputId, value, configuration = NULL)
}
\description{
Update a multiple_select input
}
