% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_button.R
\name{update_toggle_button}
\alias{update_toggle_button}
\title{Update a toggle_button input}
\usage{
update_toggle_button(session, inputId, value, configuration = NULL)
}
\description{
Update a toggle_button input
}
